Soft time scaling

+++ src/utils.rs
@@ -157,9 +157,9 @@ pub fn str_to_move(mov: &str) -> Move {
     [from, to]
 }
 
-pub fn milliseconds_elapsed(start_time: Instant) -> u32 {
+pub fn milliseconds_elapsed(start_time: Instant) -> u64 {
     let now = Instant::now();
-    now.duration_since(start_time).as_millis() as u32
+    now.duration_since(start_time).as_millis() as u64
 }
 
 pub fn incremental_sort(moves: &mut MovesArray, num_moves: u8, moves_scores: &mut [u8; 256], i: usize) -> (Move, u8)
